# docs/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "version": "1.0.0",
    "title": "AquaTrack API",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "description": "The AquaTrack API allows users to log their daily water intake, edit or delete records, and track their progress towards a daily hydration goal. Users can view their current day's water consumption and analyze their intake over the month."
  },
  "servers": [
    {
      "url": "https://aqua-track-app.onrender.com"
    },
    {
      "url": "http://localhost:3000"
    }
  ],
...
  "paths": {
...
    "/water/{id}": {
      "patch": {
        "tags": [
          "Water"
        ],
        "summary": "Edit water intake",
        "description": "Updates an existing water intake record.",
        "operationId": "updateWaterRecord",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string",
              "example": "67cf23f02f0a65716cfc2002"
            },
            "required": true,
            "description": "ID of the water record to update"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "date": {
                    "type": "string",
                    "pattern": "^(\\d{4})-(\\d{2})-(\\d{2})$",
                    "description": "Date of the water record in the format yyyy-mm-dd",
                    "example": "2025-03-08"
                  },
                  "volume": {
                    "type": "integer",
                    "description": "Volume of water (in ml)",
                    "example": 1500,
                    "minimum": 50,
                    "maximum": 5000
                  },
                  "time": {
                    "type": "string",
                    "pattern": "^([01]?[0-9]|2[0-3]):([0-5]?[0-9])$",
                    "description": "Time of the water record in the format hh:mm",
                    "example": "10:30"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully updated water record",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully updated water record!"
                    },
                    "data": {
                      "type": "object",
                      "$ref": "#/components/schemas/water"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      },
      "delete": {
        "tags": [
          "Water"
        ],
        "summary": "Delete a water record",
        "operationId": "deleteWaterRecord",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string",
              "example": "67cf23f02f0a65716cfc2002"
            },
            "required": true,
            "description": "ID of the water record to delete"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
    "/water/today": {
      "get": {
        "tags": [
          "Water"
        ],
        "summary": "Today's water intake",
        "operationId": "getWaterToday",
        "description": "Retrieves all water intake records for the current day for a logged-in user.\n",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "required": true,
            "description": "The date for which to fetch water consumption data.",
            "schema": {
              "type": "string",
              "format": "date",
              "example": "2025-03-11"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK! - Everything worked as expected",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Water consumption data retrieved successfully"
                    },
                    "data": {
                      "type": "object",
                      "required": [
                        "data"
                      ],
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/water"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
...
      "water": {
        "type": "object",
        "required": [
          "date",
          "time",
          "volume"
        ],
        "properties": {
          "_id": {
            "type": "string",
            "description": "Water record ID",
            "example": "67cf23f02f0a65716cfc2002"
          },
          "userId": {
            "type": "string",
            "description": "User ID",
            "example": "67cf170767c430dab2ce8d79"
          },
          "date": {
            "type": "string",
            "pattern": "^(\\d{4})-(\\d{2})-(\\d{2})$",
            "description": "Date of the water record in the format yyyy-mm-dd",
            "example": "2024-12-31"
          },
          "time": {
            "type": "string",
            "pattern": "^([01]?[0-9]|2[0-3]):([0-5]?[0-9])$",
            "description": "Time of the water record in the format hh:mm",
            "example": "23:59"
          },
          "volume": {
            "type": "integer",
            "description": "Volume of water (in ml)",
            "example": 650,
            "minimum": 50,
            "maximum": 5000
          }
        }
      }
    },
    "responses": {
      "400": {
        "description": "Bad request",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 400
                },
                "message": {
                  "type": "string",
                  "example": "Missing or invalid data"
                },
                "data": {
                  "type": "object",
                  "oneOf": [
                    {
                      "type": "null"
                    },
                    {
                      "type": "object"
                    }
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Invalid request parameters."
                    }
                  }
                }
              }
            }
          }
        }
      },
      "401": {
        "description": "Unauthorized error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message",
                "data"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 401
                },
                "message": {
                  "type": "string",
                  "example": "UnauthorizedError"
                },
                "data": {
                  "type": "object",
                  "oneOf": [
                    {
                      "type": "null"
                    },
                    {
                      "type": "object"
                    }
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Invalid or expired access token."
                    }
                  }
                }
              }
            }
          }
        }
      },
      "404": {
        "description": "NotFoundError",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 404
                },
                "message": {
                  "type": "string",
                  "example": "NotFoundError"
                },
                "data": {
                  "type": "object",
                  "oneOf": [
                    {
                      "type": "null"
                    },
                    {
                      "type": "object"
                    }
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "The requested resource does not exist."
                    }
                  }
                }
              }
            }
          }
        }
      },
...
      "500": {
        "description": "Internal Server Error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 500
                },
                "message": {
                  "type": "string",
                  "example": "Internal Server Error"
                },
                "data": {
                  "type": "object",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Something went wrong"
                    }
                  }
                }
              }
            }
          }
        }
      },
...
    }
  }
}